debug

+++ dl_helper/transforms/binctabl.py
@@ -17,7 +17,7 @@ class transform():
         self.batch_size = param.batch_size
 
         # 用于random_mask_row
-        self.rand_cols = torch.zeros(self.batch_size, self.time_length, dtype=torch.long, device=self.device)
+        self.rand_cols = torch.zeros(self.batch_size, self.time_length, device=self.device)
         for i in range(self.batch_size):
             self.rand_cols[i] = torch.sort(torch.randperm(self.raw_time_length, device=self.device)[:self.time_length])[0]
         self.rand_cols = self.rand_cols.unsqueeze(1).expand(-1, self.num_rows, -1)
